+++ src/modules/bank-account/application/use-cases/delete-bank-account.use-case.ts
@@ -1,7 +1,10 @@
 import { Injectable, Inject, ForbiddenException } from '@nestjs/common';
 import { BANK_ACCOUNT_TOKENS } from '../../domain/constants';
 import { IBankAccountRepository } from '../../domain/interfaces';
-import { BankAccountNotFoundException } from '../../domain/exceptions';
+import {
+  BankAccountNotFoundException,
+  BankAccountHasSettlementsException,
+} from '../../domain/exceptions';
 
 @Injectable()
 export class DeleteBankAccountUseCase {
@@ -22,6 +25,12 @@ export class DeleteBankAccountUseCase {
       throw new ForbiddenException(
         'You do not have access to this bank account',
       );
+    }
+
+    // Check if bank account has been used for settlements
+    const hasSettlements = await this.bankAccountRepository.hasSettlements(id);
+    if (hasSettlements) {
+      throw new BankAccountHasSettlementsException(id);
     }
 
     await this.bankAccountRepository.delete(id);
